+++ src/Addons/34thPRC_Overrides/config.cpp
@@ -19,6 +19,7 @@ class CfgPatches
 			"ace_huntir", // ACE
 			"19thMD_Vests_Kelp", // UNSC Foundries
 			"PhoenixSystems_Exosuits", // E.P.S.M ExoMod Remastered
+			"NSWep_Weapons", //UNSC Naval Special Weapons
 		};
 	};
 };
@@ -33,4 +34,19 @@ class CfgWeapons
 class CfgMagazines
 {
 	#include "data\config_magazines.hpp"
+};
+class CfgMagazineWells
+{
+	class UNSC_145x114_4rnd
+	{
+		19_UNSC_Weapons[]=
+		{
+			"4Rnd_145x114_srs99_apfsds",
+			"4Rnd_145x114_srs99_hvap",
+			"4Rnd_145x114_srs99_heap",
+			"4Rnd_145x114_srs99_apfsds_tracer",
+			"4Rnd_145x114_srs99_hvap_tracer",
+			"4Rnd_145x114_srs99_heap_tracer"
+		};
+	};
 };