+++ src/Addons/34thPRC_Overrides/data/weapons/srs99am/config_weapons.hpp
@@ -1,0 +1,17 @@
+class srifle_DMR_07_blk_F;
+class NSWep_SRS99AM: srifle_DMR_07_blk_F
+{
+	magazines[]+=
+	{
+		"4Rnd_145x114_srs99_apfsds",
+		"4Rnd_145x114_srs99_hvap",
+		"4Rnd_145x114_srs99_heap",
+		"4Rnd_145x114_srs99_apfsds_tracer",
+		"4Rnd_145x114_srs99_hvap_tracer",
+		"4Rnd_145x114_srs99_heap_tracer"
+	};
+	magazineWell[]=
+	{
+		"UNSC_145x114_4rnd"
+	};
+};