+++ src/changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - Fix shoulderless marine uniforms having less carry capacity
 - Gungnir helmet camos for Marines/ODSTS
+- Naval Special Weapons SRS99AM now can use 19th Fleet SRS Ammo
 
 ## 1.19.0
 ### Added